+++ lib/entidade/usuario.dart
@@ -8,12 +8,13 @@ class UsuarioFields {
   static final String email = "email";
   static final String tipoUsuario = "tipoUsuario";
   static final String senha = 'senha';
+
 }
 
 final String tableUsuario = 'usuario';
 
 class Usuario {
-  final Long? id;
+  final int? id;
   final String? nome;
   final String? email;
   final String? senha;
@@ -35,14 +36,14 @@ class Usuario {
       };
 
   static Usuario fromJson(Map<String, Object?> json) => Usuario(
-        id: json[UsuarioFields.id] as Long,
+        id: json[UsuarioFields.id] as int,
         nome: json[UsuarioFields.nome] as String,
         email: json[UsuarioFields.email] as String,
         tipoUsuario: json[UsuarioFields.tipoUsuario] as int,
       );
 
   Usuario copy({
-    Long? id,
+    int? id,
     String? nome,
     String? email,
     int? tipoUsuario,
