+++ lib/portas/saida/i_turma.dart
@@ -4,7 +4,7 @@ import 'package:diario_de_classe/dto/turmaDTO.dart';
 import 'package:diario_de_classe/dto/turma_alunoDTO.dart';
 
 abstract class ITurma {
-  Long salvarTurma(TurmaDTO turmaDTO);
+  int salvarTurma(TurmaDTO turmaDTO);
   
-  Long salvarTurmaAluno(TurmaAlunoDTO turmaAlunoDTO);
+  int salvarTurmaAluno(TurmaAlunoDTO turmaAlunoDTO);
 }
